+++ module/documents/item.mjs
@@ -3,19 +3,16 @@
  * @extends {Item}
  */
 export class ACItem extends Item {
-    /**
-     * Augment the basic Item data model with additional dynamic data.
-     */
-    prepareData() {
-        // As with the actor class, items are documents that can have their data
-        // preparation methods overridden (such as prepareBaseData()).
+    
+    /** @override */
+    prepareData() {       
         super.prepareData();
     }
 
     async _preCreate(data, options, user) {
         await super._preCreate(data, options, user);
         let ico = `systems/ac2d20/assets/doc-icons/${data.type}.svg`;
-        this.data.update({ 'img': ico });
+        this.updateSource({ 'img': ico });
     }
 
     /**
@@ -26,28 +23,27 @@ export class ACItem extends Item {
         // If present, return the actor's roll data.
         if (!this.actor) return null;
         const rollData = this.actor.getRollData();
-        rollData.item = foundry.utils.deepClone(this.data.data);
+        rollData.item = foundry.utils.deepClone(this.system);
 
         return rollData;
     }
 
     // FOCUS
     async addFocus() {
-        let focuses = this.data.data.focuses;
+        let focuses = this.system.focuses;
         const focus = { title: '', isfocus: false, description: '' }
         focuses = [...focuses, focus];
-        let updatedItem = { _id: this.id, data: { focuses: focuses } };
+        let updatedItem = { _id: this.id, "system.focuses": focuses  };
         await this.update(updatedItem);
     }
     async deleteFocus(_index) {
-        console.log(_index)
-        let focuses = this.data.data.focuses;
+        let focuses = this.system.focuses;
         focuses.splice(_index, 1);
-        let updatedItem = { _id: this.id, data: { focuses: focuses } };
+        let updatedItem = { _id: this.id, "system.focuses": focuses } ;
         await this.update(updatedItem);
     }
     async updateFocuses(_focuses) {
-        let updatedItem = { _id: this.id, data: { focuses: _focuses } };
+        let updatedItem = { _id: this.id, "system.focuses": _focuses };
         await this.update(updatedItem);
     }
     /**
@@ -56,20 +52,18 @@ export class ACItem extends Item {
      * @private
      */
     async roll() {
-        const item = this.data;
-
-        // Initialize chat data.
+        // Initialize chat data
         const speaker = ChatMessage.getSpeaker({ actor: this.actor });
         const rollMode = game.settings.get('core', 'rollMode');
-        const label = `[${item.type}] ${item.name}`;
+        const label = `[${this.type}] ${this.name}`;
 
         // If there's no roll data, send a chat message.
-        if (!this.data.data.formula) {
+        if (!this.system.formula) {
             ChatMessage.create({
                 speaker: speaker,
                 rollMode: rollMode,
                 flavor: label,
-                content: item.data.description ?? ''
+                content: this.system.description ?? ''
             });
         }
         // Otherwise, create a roll and send a chat message from it.
@@ -89,18 +83,21 @@ export class ACItem extends Item {
     }
 
     async sendToChat() {
-        const itemData = duplicate(this.data);
-        //console.warn(itemData)
-        itemData.isPhysical = itemData.data.hasOwnProperty('weight')
-        itemData.isWeapon = itemData.type === "weapon";
-        itemData.isArmor = itemData.type === "armor";
-        itemData.isTalent = itemData.type === "talent";
-        itemData.isSpell = itemData.type === "spell";
-        itemData.isSkillkit = itemData.type === "skillkit";
-        itemData.isSkillkit = itemData.type === "skillkit";
-        itemData.isEquipment = itemData.type === "equipment";
-        itemData.isSpecial_rule = itemData.type === "special_rule";
-        itemData.isSkill = itemData.type === "skill";
+        const itemData = duplicate(this.system);
+        itemData._id = this._id;
+        itemData.img = this.img;
+        itemData.name = this.name;
+        itemData.type = this.type;
+        itemData.isPhysical = this.system.hasOwnProperty('weight')
+        itemData.isWeapon = this.type === "weapon";
+        itemData.isArmor = this.type === "armor";
+        itemData.isTalent = this.type === "talent";
+        itemData.isSpell = this.type === "spell";
+        itemData.isSkillkit = this.type === "skillkit";
+        itemData.isSkillkit = this.type === "skillkit";
+        itemData.isEquipment = this.type === "equipment";
+        itemData.isSpecial_rule = this.type === "special_rule";
+        itemData.isSkill = this.type === "skill";
         const html = await renderTemplate("systems/ac2d20/templates/chat/item.html", itemData);
         const chatData = {
             user: game.user.id,